Move socket to it's own module
